Fix net-snmp test cases segfault
Upgrade net-snmp from 5.7.1 to 5.7.2.
5.7.2 fixes all its regression tests, snmptrapd won't crash any more when
running regression tests. With 5.7.1 it will crash no matter CGL or std
kernel if target is busy when run the regression tests. Didn't find out
root cause but it is a timing issue, with more delay in tests the test
passed and no crash.

Add net-snmp-tests subpackage.

Signed-off-by: Yao Zhao <yao.zhao@windriver.com>
Signed-off-by: Joe MacDonald <joe.macdonald@windriver.com>